services/embed-service/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
pydantic==2.10.3
python-dotenv==1.1.1
structlog==25.4.0
openai==1.59.5
numpy==1.26.4
Pillow==11.0.0
python-multipart==0.0.20
